requirements.txt
aiogram==3.12.0
aiogram-dialog==2.1.0
aiohttp==3.10.5
asyncpg
psycopg2
proto-plus
pytz
protobuf==3.20.1
sentence-transformers==3.0.1
sentencepiece==0.1.99
gigachain==0.2.14
gigachain-chroma==0.1.3
gigachain-cli==0.0.30
gigachain-community==0.2.12
gigachain-core==0.2.35.post2
gigachain-experimental==0.0.64
gigachain-text-splitters==0.2.3
gigaserve==0.2.2
python-dotenv==1.0.1
python-telegram-bot==21.4
lxml==4.9.4
qdrant-client==1.11.1
langchain-huggingface==0.0.3
langchain-postgres==0.0.9
langchain-qdrant==0.1.3
lark
grpcio==1.66.1
grpcio-status
grpcio-tools
rank-bm25==0.2.2
pydantic==2.8.2
pandas
numpy
psycopg-binary==3.2.1
psycopg-pool==3.2.2
wraps
SpeechRecognition
soundfile
telegram
ffprobe-python
pydub
regex
django-environ
aiopg